+++ packages/spright-components/src/chat/conversation/template.ts
@@ -5,6 +5,9 @@ export const template = html<ChatConversation>`
 <div class="toolbar ${x => (x.toolbarEmpty ? 'toolbar-empty' : '')}">
     <slot name="toolbar" ${slotted({ property: 'slottedToolbarElements' })}></slot>
 </div>
+<div class="start ${x => (x.startEmpty ? 'start-empty' : '')}">
+    <slot name="start" ${slotted({ property: 'slottedStartElements' })}></slot>
+</div>
 <div class="messages"><slot></slot></div>
 <div class="input ${x => (x.inputEmpty ? 'input-empty' : '')}">
     <slot name="input" ${slotted({ property: 'slottedInputElements' })}>