+++ packages/spright-components/src/chat/input/index.ts
@@ -37,6 +37,9 @@ export class ChatInput extends mixinErrorPattern(FoundationElement) {
     @attr({ attribute: 'processing', mode: 'boolean' })
     public processing = false;
 
+    @attr({ attribute: 'send-disabled', mode: 'boolean' })
+    public sendDisabled = false;
+
     /**
      * @internal
      */
@@ -44,10 +47,11 @@ export class ChatInput extends mixinErrorPattern(FoundationElement) {
     public textArea?: HTMLTextAreaElement;
 
     /**
+     * Tracks whether the send button should be disabled based on input value
      * @internal
      */
     @observable
-    public disableSendButton = true;
+    public isInputEmpty = true;
 
     /**
      * The width of the vertical scrollbar, if displayed.
@@ -78,7 +82,7 @@ export class ChatInput extends mixinErrorPattern(FoundationElement) {
      */
     public textAreaInputHandler(): void {
         this.value = this.textArea!.value;
-        this.disableSendButton = this.shouldDisableSendButton();
+        this.isInputEmpty = this.shouldDisableSendButton();
         this.queueUpdateScrollbarWidth();
     }
 
@@ -102,7 +106,7 @@ export class ChatInput extends mixinErrorPattern(FoundationElement) {
     public valueChanged(): void {
         if (this.textArea) {
             this.textArea.value = this.value;
-            this.disableSendButton = this.shouldDisableSendButton();
+            this.isInputEmpty = this.shouldDisableSendButton();
             this.queueUpdateScrollbarWidth();
         }
     }
@@ -113,7 +117,7 @@ export class ChatInput extends mixinErrorPattern(FoundationElement) {
     public override connectedCallback(): void {
         super.connectedCallback();
         this.textArea!.value = this.value;
-        this.disableSendButton = this.shouldDisableSendButton();
+        this.isInputEmpty = this.shouldDisableSendButton();
         this.resizeObserver = new ResizeObserver(() => this.onResize());
         this.resizeObserver.observe(this);
     }
@@ -157,7 +161,7 @@ export class ChatInput extends mixinErrorPattern(FoundationElement) {
 
     private resetInput(): void {
         this.value = '';
-        this.disableSendButton = true;
+        this.isInputEmpty = true;
         if (this.textArea) {
             this.textArea.value = '';
             this.textArea.focus();