+++ packages/spright-components/src/icons/all-icons.ts
@@ -1,0 +1,5 @@
+export { IconWorkItemCalendarWeek } from './work-item-calendar-week';
+export { IconWorkItemCalipers } from './work-item-calipers';
+export { IconWorkItemForklift } from './work-item-forklift';
+export { IconWorkItemRectangleCheckLines } from './work-item-rectangle-check-lines';
+export { IconWorkItemWrenchHammer } from './work-item-wrench-hammer';