+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @ni/nimble-components
 
-<!-- This log was last generated on Mon, 02 Mar 2026 00:16:21 GMT and should not be manually modified. -->
+<!-- This log was last generated on Mon, 02 Mar 2026 18:11:33 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 35.4.0
+
+Mon, 02 Mar 2026 18:11:33 GMT
+
+### Minor changes
+
+- Add theme aware token for splitter background ([ni/nimble@8c04925](https://github.com/ni/nimble/commit/8c04925fec84772102d7fbd07482e67081b029b5))
+- Bump @ni/nimble-tokens to v8.16.0
+
+## 35.3.4
+
+Mon, 02 Mar 2026 17:43:16 GMT
+
+### Patches
+
+- step and anchor-step ([ni/nimble@7b528b9](https://github.com/ni/nimble/commit/7b528b96c6462eb69faf6c9ae10986782ca353f7))
 
 ## 35.3.3